refactor: decompose settings reducer workflows
